converted isbn to long/bigint value due to int being unable to hold more than 9 digits. Also converted  phone#'s and fax's numbers in publishers and staff to String/varchar20 for the same reason. Also populated database with remaining missing data from the previous application

+++ src/main/java/beans/Book.java
@@ -11,16 +11,16 @@ package beans;
  */
 public class Book {
     
-    private int isbn;
+    private long isbn;
     private int publisherId;
     private String text;
     private String author;
 
-    public int getIsbn() {
+    public long getIsbn() {
         return isbn;
     }
 
-    public void setIsbn(int isbn) {
+    public void setIsbn(long isbn) {
         this.isbn = isbn;
     }
